+++ plugins/sonarqube/pkg/instance/instance.go
@@ -9,12 +9,6 @@ import (
 	"strings"
 
 	"github.com/kobsio/kobs/pkg/api/middleware/roundtripper"
-
-	"github.com/sirupsen/logrus"
-)
-
-var (
-	log = logrus.WithFields(logrus.Fields{"package": "sonarqube"})
 )
 
 // Config is the structure of the configuration for a single Opsgenie instance.