+++ plugins/users/users.go
@@ -9,18 +9,15 @@ import (
 	"github.com/kobsio/kobs/pkg/api/clusters"
 	"github.com/kobsio/kobs/pkg/api/middleware/errresponse"
 	"github.com/kobsio/kobs/pkg/api/plugins/plugin"
+	"github.com/kobsio/kobs/pkg/log"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/render"
-	"github.com/sirupsen/logrus"
+	"go.uber.org/zap"
 )
 
 // Route is the route under which the plugin should be registered in our router for the rest api.
 const Route = "/users"
-
-var (
-	log = logrus.WithFields(logrus.Fields{"package": "users"})
-)
 
 // Config is the structure of the configuration for the users plugin.
 type Config struct{}
@@ -59,13 +56,14 @@ func isMember(teams []user.TeamReference, defaultCluster, defaultNamespace, clus
 // getUsers returns a list of users for all clusters and namespaces. We always return all users for all clusters and
 // namespaces. For this we are looping though the loaded clusters and callend the GetUsers function for each one.
 func (router *Router) getUsers(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("getUsers")
+	log.Debug(r.Context(), "Get users.")
 
 	var users []user.UserSpec
 
 	for _, cluster := range router.clusters.Clusters {
 		user, err := cluster.GetUsers(r.Context(), "")
 		if err != nil {
+			log.Error(r.Context(), "Could not get users.")
 			errresponse.Render(w, r, err, http.StatusBadRequest, "Could not get users")
 			return
 		}
@@ -73,7 +71,7 @@ func (router *Router) getUsers(w http.ResponseWriter, r *http.Request) {
 		users = append(users, user...)
 	}
 
-	log.WithFields(logrus.Fields{"count": len(users)}).Tracef("getUsers")
+	log.Debug(r.Context(), "Get users result.", zap.Int("usersCount", len(users)))
 	render.JSON(w, r, users)
 }
 
@@ -85,16 +83,18 @@ func (router *Router) getUser(w http.ResponseWriter, r *http.Request) {
 	namespace := r.URL.Query().Get("namespace")
 	name := r.URL.Query().Get("name")
 
-	log.WithFields(logrus.Fields{"cluster": clusterName, "namespace": namespace, "name": name}).Tracef("getUser")
+	log.Debug(r.Context(), "Get User parameters.", zap.String("cluster", clusterName), zap.String("namespace", namespace), zap.String("name", name))
 
 	cluster := router.clusters.GetCluster(clusterName)
 	if cluster == nil {
+		log.Error(r.Context(), "Invalid cluster name.", zap.String("cluster", clusterName))
 		errresponse.Render(w, r, nil, http.StatusBadRequest, "Invalid cluster name")
 		return
 	}
 
 	user, err := cluster.GetUser(r.Context(), namespace, name)
 	if err != nil {
+		log.Error(r.Context(), "Could not get user.", zap.Error(err))
 		errresponse.Render(w, r, err, http.StatusBadRequest, "Could not get user")
 		return
 	}
@@ -107,10 +107,13 @@ func (router *Router) getTeams(w http.ResponseWriter, r *http.Request) {
 	defaultClusterName := r.URL.Query().Get("cluster")
 	defaultNamespace := r.URL.Query().Get("namespace")
 
+	log.Debug(r.Context(), "Get User parameters.", zap.String("defaultCluster", defaultClusterName), zap.String("defaultNamespace", defaultNamespace))
+
 	var data getTeamsData
 
 	err := json.NewDecoder(r.Body).Decode(&data)
 	if err != nil {
+		log.Error(r.Context(), "could not decode request body.", zap.Error(err))
 		errresponse.Render(w, r, err, http.StatusBadRequest, "Could not decode request body")
 		return
 	}
@@ -130,12 +133,14 @@ func (router *Router) getTeams(w http.ResponseWriter, r *http.Request) {
 
 		cluster := router.clusters.GetCluster(c)
 		if cluster == nil {
+			log.Error(r.Context(), "Invalid cluster name.", zap.String("cluster", c))
 			errresponse.Render(w, r, nil, http.StatusBadRequest, "Invalid cluster name")
 			return
 		}
 
 		t, err := cluster.GetTeam(r.Context(), n, team.Name)
 		if err != nil {
+			log.Error(r.Context(), "Could not get team.", zap.String("team", team.Name))
 			errresponse.Render(w, r, err, http.StatusBadRequest, "Could not get team")
 			return
 		}
@@ -152,7 +157,7 @@ func (router *Router) getTeam(w http.ResponseWriter, r *http.Request) {
 	teamNamespace := r.URL.Query().Get("namespace")
 	teamName := r.URL.Query().Get("name")
 
-	log.WithFields(logrus.Fields{"cluster": teamCluster, "namespace": teamNamespace, "name": teamName}).Tracef("getTeam")
+	log.Debug(r.Context(), "Get team parameters.", zap.String("cluster", teamCluster), zap.String("namespace", teamNamespace), zap.String("name", teamName))
 
 	var users []user.UserSpec
 	var filteredUsers []user.UserSpec
@@ -160,6 +165,7 @@ func (router *Router) getTeam(w http.ResponseWriter, r *http.Request) {
 	for _, cluster := range router.clusters.Clusters {
 		user, err := cluster.GetUsers(r.Context(), "")
 		if err != nil {
+			log.Error(r.Context(), "Could not get users.")
 			errresponse.Render(w, r, err, http.StatusBadRequest, "Could not get users")
 			return
 		}